requirements.txt
langchain==0.1.20
langchain-openai==0.1.7
openai==1.26.0
python-dotenv==1.0.1
chromadb==0.5.2
